scraper/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.2
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.3
comm==0.2.1
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
executing==2.0.1
fastapi==0.88.0
filelock==3.13.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.38.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
multidict==6.0.4
mysql-connector-python==8.0.32
nest-asyncio==1.6.0
numpy==2.2.2
openai==0.26.4
outcome==1.3.0.post0
packaging==23.0
pandas==2.2.3
parso==0.8.3
patsy==0.5.3
pexpect==4.9.0
Pillow==9.4.0
platformdirs==4.2.0
plotly==5.13.0
plotly-express==0.4.1
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.21.0
pydantic==1.10.2
pygame==2.6.1
Pygments==2.14.0
pynecone==0.1.14
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==25.1.2
redis==4.4.2
requests==2.32.3
rfc3986==1.5.0
rich==12.6.0
scikit-learn==1.4.1.post1
scipy==1.10.0
seaborn==0.13.2
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.29
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
stack-data==0.6.3
starlette==0.22.0
statsmodels==0.13.5
tenacity==8.1.0
threadpoolctl==3.3.0
tk==0.1.0
tornado==6.4
tqdm==4.64.1
traitlets==5.14.1
trio==0.28.0
trio-websocket==0.11.1
typer==0.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.20.0
virtualenv==20.25.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.8.2
